Split mixed_extension unittest
